Enable 64bit builds on Windows.
* Added runcmake64.bat for 64bit configuration (creates bin64 directory)
* Extended boost build script for 64bit binaries
* Fixed warnings failing the build in 64bit mode

+++ src/gui/ChessSet.cpp
@@ -39,7 +39,7 @@ ChessSet::~ChessSet() {
 }
 
 int ChessSet::getResourcesCount() {
-	return m_extResources.size() + 1;	// +1 for the board
+	return static_cast<int>(m_extResources.size()) + 1;	// +1 for the board
 }
 
 void ChessSet::loadResources() {
